Move the control protocol enum to main client lib Side benefit: forcing documentation
